Clear many build warnings

+++ api/oc_helpers.c
@@ -123,9 +123,8 @@ bool
 _oc_string_array_add_item(oc_string_array_t *ocstringarray, const char str[])
 {
   bool success = false;
-  int i;
-  for (i = 0; i < (int)oc_string_array_get_allocated_size(*ocstringarray);
-       i++) {
+  size_t i;
+  for (i = 0; i < oc_string_array_get_allocated_size(*ocstringarray); i++) {
     if (oc_string_array_get_item_size(*ocstringarray, i) == 0) {
       success = oc_string_array_set_item(*ocstringarray, str, i);
       break;
@@ -138,7 +137,7 @@ void
 oc_join_string_array(oc_string_array_t *ocstringarray, oc_string_t *ocstring)
 {
   size_t len = 0;
-  int i;
+  size_t i;
   for (i = 0; i < oc_string_array_get_allocated_size(*ocstringarray); i++) {
     const char *item =
       (const char *)oc_string_array_get_item(*ocstringarray, i);
